requirements.txt
discord
discord.py
pip
praw
python-dotenv
requests
async
wavelink
yt-dlp